refactor: share flat update helpers
