+++ src/optimizers/ExtendedKalmanFilter.py
@@ -1,13 +1,13 @@
 import torch
 
-from ._utils import add_flat_update_
+from ._utils import _FlatUpdateOptimizerMixin
 from ._utils import kalman_update
 from ._utils import residual_jacobian
 from ._utils import residual_sum_squares
 from ._utils import trainable_params
 
 
-class ExtendedKalmanFilter(torch.optim.Optimizer):
+class ExtendedKalmanFilter(_FlatUpdateOptimizerMixin, torch.optim.Optimizer):
     """Extended Kalman filter optimizer for residual-vector closures.
 
     This optimizer assumes a single parameter group and expects `closure()` to
@@ -89,10 +89,6 @@ class ExtendedKalmanFilter(torch.optim.Optimizer):
     def loss(self, errors):
         return residual_sum_squares(errors)
     
-    @torch.no_grad()
-    def update_weights(self, updates):
-        add_flat_update_(trainable_params(self.param_groups), updates)
-
     def step(self, closure = None):
 
         assert len(self.param_groups) == 1
